Ensure that unlabeling of PRs is ignored.
This ensures that checks to see if a PR is mergable is not carried out.
Whilst VersionBot won't attempt to create a new merge, it will complain that
it can't. This means that scripted labeling/unlabeling of a PR will have an
error attached at the end of a successful merge. We need to quash this.

Also lower number of retries to a total of 3.

Connects to #37

Change-Type: patch