Improve status checks for required external services; reduce verbose 'spamming' of PR on status failures.
Previously, VersionBot was very verbose in its status checks; either everything succeeded
or there was a failure. This did not work well with pending statuses. More than that, it
made comments which weren't actually necessary as the Github status check dialogue makes
it obvious why a PR can't be merged.

Additionally, there are some services which actually don't even report a status such as
Jenkins which simply ignores PRs until it's polled and is ready for them. This has the
effect of Github not being able to return a status, or even an entry, for Jenkins as it
hasn't been acknowledged by it. We therefore still attempt to go through with a merge,
but don't report if the service is still outstanding. Jenkins eventually catches up, does
a build, reports a status and the PR is merged.

Connects to #62

Change-type: patch