ProcBots: Add ability to retrieve underlying SDK API instance object/handle for a service.
Previously, should a service use an underlying SDK (such as a Node module)
to communicate externally with the service they provide access for, the
API instance would have to be communicated separately (in the case of VersionBot,
it used the API instance sent through cooked events). Whilst not used directly
for emitting/listening, there are instances where this is useful (for example to
form commands for an emitter, such as relevant method names and parameters).

This commit now allows Listeners/Emitters to expose the underlying SDK API instance
(should one exist) to allow bots to retrieve them more easily. It should be noted
that bots should *not* use the APIs directly to communicate with external services,
this should still occur via emitters.

Updates all current services/bots as required.

Connects to #94

Change-type: patch