Ensure that only the `master` branch is used for document sources.
Currently whenever we commit new code, the entire `docs` directory gets
regenerated with links to new commit blobs, even if the actual documentation
hasn't changed. This becomes extremely noisy.

By ensuring that the `master` branch is always used, we ensure that the only
document changes are those of the code that's changed.

Connects to #102

Change-type: patch