Add autogeneration of Waffleboard commands should the `Connects-To: #<number>` tag be present.
This ensures that Waffleboard statuses can be updated whilst still ensuring that
we can use a standard `tag-name:` standard for all our commits.
When a `Connects-To: #<number>` tag is found in either the PR description or *any*
commits for the PR, the description is updated to include autogenerated lines
that allow Waffleboard to update itself based on them.

Change-Type: minor