Improve `ProcBot:dispatchToEmitter()` method to simplify calling/return conventions.
Previously a Client Bot had to construct a relevant ServiceEmitRequest before
calling `dispatchToEmitter()`, and then had to check the ServiceEmitResponse
on its fulfilment. This is overkill considering that a specifically named
ServiceEmitter is being invoked. Now, `dispatchToEmitter()` takes any relevant
data structure for the named emitter and, on response from the service, either
throws any returned error or fulfils with the data that was returned. This
simplifies the calling in a client, which may now just do:

this.dispatchToEmitter(<serviceEmitterName>, <serviceDataStructure>)

Change-Type: major