ProcBots: Modify configuration file loading so that services themselves know how to retrieve them.
Previously, bots themselves had to make emitter calls to retrieve configuration files
from a particular service. This is now changed so that a call to a particular service will
retrieve a configuration file instead. This makes much more sense as it means that a
bot doesn't have to know how to do so, just the location of the configuration file.

Also fixes issue with gulp and documentation building.

Change-type: minor