VersionBot: Add ability to specify required tags for a PR in a repository configuration file.
This functionality allows repository maintainers to determine which footer tags in
a PR are mandatory or optional.
Tags keys are considered case insensitive, and optionally may define how many
occurences should be present in all of the commits for a PR, as well as defining
valid values.

The `change-type` tag is *always* assumed to be present, regardless of other
tag definitions. Should the configuration file redefine the `change-type` tag,
then that definition will override the inbuilt one.

The following shows how to format required tag definitions in the `repository.yml`
configuration:

```
required_tags:
    <tagname>:
        occurrence: 'all' | 'once' | 'never'
        values: <regexp body>
	flags: <valid flags>
```

Both the `occurrence` and `values` properties are optional, if present:
* `occurrence`[optional] - Determines how many occurrences of the tag should occur within a
                           PR. Valued values are:
			   * `once` - The tag should occur in at least one commit in the PR
			   * `all` - The tag should occur in every commit in the PR
			   * `never` - The tag should never occur in any commit in the PR
* `values`[optional] - A regular expression body defining a tag's valid values.
  	               This body should be defined to match values that occur after a tag prefix, and
		       should exclude the leading and trailing solidus from the expression. Bodies
		       will always be suffixed with a `$` when compiled, which ensures that a value
		       can only exist on a single line.
		       Defaults to `.*` should it not be defined.
* `flags`[optional] - Only valid (and parsed) should the `values` property be present.
                      This is a string consisting of the regular expression flags that
                      pertain to the regular expression body defined in the `values`
                      property. Valid values are: `i`, `u` and `y`. The `g` and `m`
                      flags have no meaningful value in the tests and will be ignored.

As an example, the following is a valid set of definitions for a repository,
and includes the `change-type` tag explicitly.

```
required-tags:
    signed-off-by:
        occurrence: all
    changelog-entry:
        occurrence: never
    change-type:
        occurrence: once
        values: \s*(patch|minor|major)\s*
        flags: i
```

Connects-to: #144

Change-type: minor