SyncBot, ProcBots: Short-list discourse threads when looking for posts
This makes the discourse API perform some of the filtering when suitable input is provided.  This code is not used in a more general capacity, and is going to be part of the Great Refactor, but fixes a problem we're experiencing now.  Threads where SyncBot joins quite late could have their `connects to` message not in the first page of results, which causes new thread creation.

* Short-list post searches in Discourse API where easy

Change-Type: patch