ProcBots: Add ability to have multiple ServiceListener/ServiceEmitter instances of the same Service per Client Bot.
This allows Clients to talk to the same Service with different authentication details, or to
a different endpoint, for example.

This is a backwards compatible feature, so ensures that all legacy Clients still work
as expected.

Have also added a variable delay mechanism into the environment CLI utilities,
and made retry for applying versionist in VersionBot delayed by 5 seconds at a time.
This should give Github enough time to update its internal token details.

Connects-to: #216

Change-type: minor