VersionBot: Ensure the `procbots/versionbot/no-checks` label correctly allows manual merging.
This commit ensures that when the `no-checks` label is applied, the `AutoMerge` and
`Versionist` statuses are automatically successful. It will *not* pass the `Reviewers`
status, however, which still needs to be valid for merging.

Additionally, a couple of bugs have been fixed to allow a move to TypeScript 2.4.x,
mostly around the use of the `Worker` class' generic type usage.

Connects-to: #227

Change-type: patch