SyncBot: Synchronise Front conversation states into Flowdock
Code the transmission of Front state to Flowdock tags. The pidgin now
has a dictionary of states to clear or set. The syncbot route function
now has a section that routes state changes. The Front translator
passes state changes along. The Flowdock translator stores state as
initial message tags.

Adjust the supporting framework to support the broader scope of events.
The types ave been updated to have a richer object definition and new
name. The line emitted to the logs falls back onto the thread title.
The metadata encoding requires the message and the metadata separately.

Tidy up a couple of details while I was there. The intercomHack
property, since it is used to route, best belongs in the service
details. The Front Session is now readonly. A regex had redundant
escape character.

Change-type: patch
Connects-to: #564
Signed-off-by: Andrew Lucas <andrewl@resin.io>