flowdockhashtag: Fix .match is not a function critical error
We found this because of this error: `TypeError: message.content.match
is not a function`.

The problem is that in various event types, `content` is not a string,
but an object. See https://www.flowdock.com/api/message-types for
examples.

We believe this ended up causing some messages to be lost when syncing
from Front to Flowdock.

Change-type: patch
Signed-off-by: Juan Cruz Viotti <juan@balena.io>