Workaround for missed Front -> FD messages
There's an assumptions stated at
https://github.com/balena-io-modules/resin-procbots/issues/301, that

>  an event inputted in Front/Intercom has an empty subject property, where an event inputted from Flowdock with a subject property retains that.

This is only half-true, as events in Front/Intercom sometimes have a
'Re: ' title that breaks the sync. We've still haven't gotten to the
bottom of when/why it happens, but this temporary fix should be enough to prevent lost
messages given that the 'Re: ' behaviour is consistent.

Change-type: patch
Signed-off-by: Kostas Lekkas <kostas@balena.io>